Feedback-request + Feedback

+++ src/main/java/ita/softserve/course_evaluation/repository/CourseRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CourseRepository extends CrudRepository<Course, Integer> {
+public interface CourseRepository extends CrudRepository<Course, Long> {
 
 }
